+++ JobBackgroundService/Job/Service/JobService.cs
@@ -8,7 +8,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 
-namespace Job
+namespace Job.Service
 {
     public class JobService : BackgroundService
     {
